2009-10-23  Brad Taylor  <brad@getcoded.net>
In MoonAtkBridge/MoonAtkBridge/:
	* Moonlight.AtkBridge/Adapter.cs: Emit accessible-name event when a
	control's name changes.

	* Moonlight.AtkBridge/AutomationBridge.cs: Add better debugging for
	event names.

In MoonAtkBridge/tests/:
	* moonlight/browser/__init__.py: Close Firefox with altF4 first, then
	if it's not dead yet, kill it.

	* moonlight/testing.py:
	* moonlight/__init__.py: Override unittest.TestCase, and wrap all
	assert* and fail* methods in retryUntilTrue.

	* test_value.py:
	* test_adapter.py:
	* test_bug537507.py:
	* test_selection.py:
	* test_rangevalue.py:
	* test_expandcollapsetoggleinvoke.py: Use new TestCase class.

	* assets/Makefile.am:
	* assets/AdapterEventsTest/App.xaml:
	* assets/AdapterEventsTest/Page.xaml:
	* assets/AdapterEventsTest/App.xaml.cs:
	* assets/AdapterEventsTest/Makefile.am:
	* assets/AdapterEventsTest/Page.xaml.cs: Add AdapterEventsTest sample.

	* test_adapter_events.py: Add test that verifies that accessible name,
	sensitivity, bounds changed and focus events are fired correctly by
	Adapter.

svn path=/trunk/uia2atk/; revision=144735

+++ MoonAtkBridge/tests/moonlight/browser/__init__.py
@@ -26,6 +26,7 @@
 import os
 import signal
 
+from time import sleep
 from strongwind import *
 
 def launchAddress(uri, browser='firefox', profile='dev', name='Namoroka'):
@@ -80,7 +81,13 @@ class FirefoxBrowser(accessibles.Application):
         self.slControl = self.mainFrame.findFiller('Silverlight Control')
 
     def kill(self):
+        self.mainFrame.altF4()
+        sleep(config.MEDIUM_DELAY)
+
         try:
-            os.killpg(self.proc.pid, signal.SIGKILL)
+            self.proc.poll()
+            if not self.proc.returncode:
+                os.killpg(self.proc.pid, signal.SIGKILL)
+                self.proc.wait()
         except:
             print "Unable to send pid %d SIGKILL..." % self.proc.pid
